Centralize creation of search icon
