+++ app/src/main/res/menu/search.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+            android:id="@+id/action_search"
+            android:icon="?attr/action_search"
+            app:showAsAction="collapseActionView|always"
+            app:actionViewClass="androidx.appcompat.widget.SearchView"
+            android:title="@string/search_label"/>
+</menu>
